# microservicio_uno/app/Models/Fire.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Fire extends Model
{
    use HasFactory;

    public function city()
    {
        $this->belongsTo('city');
    }
}